+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.dao.entity.SectionEntity;
 import wooteco.subway.domain.Section;
 
 @Repository
