+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.dao.entity.StationEntity;
 import wooteco.subway.domain.Station;
 
 @Repository
