+++ src/main/java/wooteco/subway/dao/entity/StationEntity.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.entity;
 
 public class StationEntity {
