+++ src/main/java/wooteco/subway/service/dto/request/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto.request;
 
 public class LineUpdateRequest {
