+++ src/main/java/wooteco/subway/service/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto.request;
 
 public class StationRequest {
