+++ src/main/java/wooteco/subway/service/dto/response/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto.response;
 
 public class StationResponse {
