+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
+import wooteco.subway.dao.entity.StationEntity;
 import wooteco.subway.domain.Station;
 
 @JdbcTest
